Feat: 레디스 트랜잭션 설정 및 회원 선택물품만 결제로 넘어가기 기능

+++ src/main/java/shop/s5g/shop/service/cart/impl/CartServiceImpl.java
@@ -30,6 +30,7 @@ import shop.s5g.shop.repository.cart.CartRedisRepository;
 import shop.s5g.shop.repository.cart.CartRepository;
 
 
+import shop.s5g.shop.repository.delivery.DeliveryFeeRepository;
 import shop.s5g.shop.service.cart.CartService;
 import shop.s5g.shop.service.member.MemberService;
 
@@ -42,7 +43,7 @@ public class CartServiceImpl implements CartService {
     private final CartRedisRepository cartRedisRepository;
     private final MemberService memberService;
     private final BookRepository bookRepository;
-
+    private final DeliveryFeeRepository deliveryFeeRepository;
 
 
     // 회원아이디를 이용해 Mysql 에 저장되어있는 Cart 리스트반환
@@ -57,10 +58,13 @@ public class CartServiceImpl implements CartService {
     // 로그인 했을 때 세션스토리지와 db에 있는 걸 합쳐서 레디스에 옮김
     @Transactional
     @Override
-    public int saveMergedCartToRedis(String customerLoginId,List<CartBookInfoRequestDto> cartBookInfoListInSession) {
+    public void saveMergedCartToRedis(String customerLoginId,List<CartBookInfoRequestDto> cartBookInfoListInSession) {
 
         // 정상로그아웃을 했을 경우에는 해당회원의 레디스 loginFlag 를 삭제한다.
         // 근데 만약 loginFlag 가 존재한다면 비정상로그아웃으로 간주하고 db에 있는 걸 안 가져오고 레디스에 세션 스토리지 내용만 추가하는 방법 고려
+        if (customerLoginId.isBlank()) {
+            throw new BadRequestException("CustomerLoginId Is Not Valid");
+        }
 
         if (Objects.isNull(getLoginFlag(customerLoginId))) {    // null 이라는 것은 정상 로드아웃이 되었다는 것
             List<Cart> booksInDb = getBooksInDbByCustomerId(customerLoginId);
@@ -79,9 +83,11 @@ public class CartServiceImpl implements CartService {
             cartRedisRepository.putBook(cartBookInfo.bookId(), cartBookInfo.quantity(),
                 customerLoginId);
         });
+    }
 
+    @Override
+    public int getCartCountInRedis(String customerLoginId){
         return cartRedisRepository.getBooksInRedisCart(customerLoginId).size();
-
     }
 
     // 로그아웃할 때 레디스에 있는 걸 db 에 저장한다.
@@ -127,6 +133,10 @@ public class CartServiceImpl implements CartService {
     @Override
     public void controlQuantity(Long bookId, int change, String customerLoginId) {
 
+        if (customerLoginId.isBlank()) {
+            throw new BadRequestException("CustomerLoginId Is Not Valid");
+        }
+
         if (change > 0) {
             putBook(bookId, 1, customerLoginId);
         } else {
@@ -138,6 +148,10 @@ public class CartServiceImpl implements CartService {
     @Override
     public void putBook(Long bookId, Integer quantity, String customerLoginId) {
 
+        if (customerLoginId.isBlank()) {
+            throw new BadRequestException("CustomerLoginId Is Not Valid");
+        }
+
         if (bookRepository.findById(bookId).isPresent()) {
             cartRedisRepository.putBook(bookId, quantity, customerLoginId);
         }else {
@@ -146,16 +160,22 @@ public class CartServiceImpl implements CartService {
     }
 
 
+    @Transactional
     @Override
-    // 필요함 근데 레디스관련이라  Transactional 필요없음
     public void reduceBookQuantity(Long bookId, String customerLoginId) {
+
         cartRedisRepository.reduceBookQuantity(bookId, customerLoginId);
     }
 
 
+    @Transactional
     @Override
-    // 필요함 근데 레디스관련이라  Transactional 필요없음
     public void deleteBookFromCart(Long bookId, String customerLoginId) {
+
+        if (customerLoginId.isBlank()) {
+            throw new BadRequestException("CustomerLoginId Is Not Valid");
+        }
+
         cartRedisRepository.deleteBookFromCart(bookId, customerLoginId);
     }
 
@@ -220,39 +240,36 @@ public class CartServiceImpl implements CartService {
 
     //TODO 이후 배달비 관련 테이블이 완성되면 배달비랑 조건 수정필요
     @Override
-    // 필요함 근데 Transactional 필요없음
     public CartDetailInfoResponseDto getTotalPriceAndDeliverFee(
         List<CartBooksResponseDto> cartBooks) {
 
         BigDecimal totalPrice = cartBooks.stream().map(cartBook -> cartBook.discountedPrice()
                 .multiply(BigDecimal.valueOf(cartBook.quantity())))
             .reduce(BigDecimal.ZERO, BigDecimal::add);
-        int deliveryFee = 3000;
-        int freeShippingThreshold = 30000;
+
+        long deliveryFee = 3000;
+        long freeShippingThreshold = 30000;
         return new CartDetailInfoResponseDto(totalPrice, deliveryFee, freeShippingThreshold);
     }
 
 
 
     @Override
-    // 필요함 근데 Transactional 필요없음
     public void setLoginFlag(String customerLoginId) {
         cartRedisRepository.setLoginFlag(customerLoginId);
     }
+
     @Override
-    // 필요함 근데 Transactional 필요없음
     public Boolean getLoginFlag(String customerLoginId) {
         return cartRedisRepository.getLoginFlag(customerLoginId);
     }
 
     @Override
-    // 필요함 근데 Transactional 필요없음
     public void deleteLoginFlag(String customerLoginId) {
         cartRedisRepository.deleteLoginFlag(customerLoginId);
     }
 
     @Override
-    // 필요함 근데 Transactional 필요없음
     public Map<Long, Integer> getBooksInRedisCart(String customerLoginId) {
         Map<Long, Integer> result = new HashMap<>();
 
@@ -265,6 +282,7 @@ public class CartServiceImpl implements CartService {
 
     }
 
+    @Override
     public Map<Long, CartFieldValue> getBooksInRedisCartWithStatus(String customerLoginId) {
         Map<Object, Object> booksInRedisCart = cartRedisRepository.getBooksInRedisCart(
             customerLoginId);
@@ -296,14 +314,16 @@ public class CartServiceImpl implements CartService {
 
 
     @Override
-    // 필요함 근데 Transactional 필요없음
     public void deleteOldCart(String customerLoginId) {
         cartRedisRepository.deleteOldCart(customerLoginId);
-
     }
 
     @Transactional
     public void removeAccount(String customerLoginId) {
+
+        if (customerLoginId.isBlank()) {
+            throw new BadRequestException("SessionId Is Not Valid");
+        }
 
         deleteOldCart(customerLoginId);
         cartRepository.deleteAllByCartPk_CustomerId(
@@ -312,8 +332,13 @@ public class CartServiceImpl implements CartService {
 
     }
 
+    @Transactional
     @Override
     public List<CartBookInfoRequestDto> getBooksWhenPurchase(String customerLoginId) {
+
+        if (customerLoginId.isBlank()) {
+            throw new BadRequestException("SessionId Is Not Valid");
+        }
 
         Map<Long, Integer> booksInRedisCart = getBooksInRedisCartWithStatusTrue(customerLoginId);
 
@@ -326,9 +351,14 @@ public class CartServiceImpl implements CartService {
         return books;
     }
 
+    @Transactional
     @Override
     public void changeBookStatus(String customerLoginId,
         CartBookSelectRequestDto cartBookSelectRequestDto) {
+
+        if (customerLoginId.isBlank()) {
+            throw new BadRequestException("SessionId Is Not Valid");
+        }
 
         cartRedisRepository.changeBookStatus(customerLoginId, cartBookSelectRequestDto);
